Performance/power work
The previous approach would constantly re-draw the viewport, even if nothing had changed that required an update. This commit revises the code to only reconstruct geometry and draw if really required (either some relevant settings were changed, or the viewport was resized).

On real world applications, this greatly reduces the power draw for an otherwise idle viewport.